tests: Add 'nova' and 'glance' OpenStack client tests
Add tests that use the standard upstream 'nova' and 'glance' python
clients. Compare the output of the different commands to pre-defined
content that is read from data files.

Note that a mock libvirt library is being used and no real instances
are spun up. So there's no dependecy on a working libvirt environment.

Eventually, we want to test every supported OpenStack client command.

Signed-off-by: Juerg Haefliger <juergh@gmail.com>