Collapse the different API servers into a single instance
This is a massive commit that basically gets rid of all the different
API server threads and instead uses just a single instance. This is now
supported in OpenStack (and specifically the clients) and simplifies
things considerably, so make use of it.

At the same time, this commit also removes the EC2 metadata server and
defaults to forcing the use of a config drive.

Lastly, update all the tests so they pass with the latest clients.

Yes, this should be broken up into a series of commits but I'm lazy :-(

Signed-off-by: Juerg Haefliger <juergh@gmail.com>